cleanup- Domains

+++ FrontEnd/AH.Web/Server/Services/DomainService.cs
@@ -7,30 +7,22 @@ namespace AH.Web.Server.Services;
 public class DomainService : IDomainService
 {
     private readonly HttpClient _httpClient;
+
     public DomainService(HttpClient httpClient, IHttpContextAccessor httpContextAccessor)
     {
         _httpClient = httpClient;
-        var token = httpContextAccessor?.HttpContext?.Request.Headers["Authorization"].ToString() ?? string.Empty;
-        _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token.Replace("Bearer ", string.Empty));
+        var token = httpContextAccessor.HttpContext?.Request.Headers["Authorization"].ToString() ?? string.Empty;
+        _httpClient.DefaultRequestHeaders.Authorization =
+            new AuthenticationHeaderValue("Bearer", token.Replace("Bearer ", string.Empty));
     }
-
-    public async Task<List<DomainResponse>> GetDomains()
-    {
-        return await _httpClient.Get<DomainResponse>("v1/domains");
-    }
+    public async Task<List<DomainResponse>> GetDomains() => await _httpClient.Get<DomainResponse>("v1/domains");
 
     public async Task<DomainResponse> UpdateDomain(DomainResponse domain)
-    {
-        return await _httpClient.Upsert<DomainResponse>($"v1/domains/{domain.UId}", domain, true);
-    }
+        => await _httpClient.Upsert<DomainResponse>($"v1/domains/{domain.UId}", domain, true);
 
-    public async Task<DomainResponse> CreateDomain(DomainResponse domain)
-    {
-        return await _httpClient.Upsert<DomainResponse>($"v1/domains/{domain.UId}", domain, false);
-    }
+    public async Task<DomainResponse> CreateDomain(DomainResponse domain) =>
+         await _httpClient.Upsert<DomainResponse>($"v1/domains/{domain.UId}", domain, false);
 
-    public async Task DeleteDomain(Guid uid)
-    {
-        await _httpClient.Delete($"v1/domains/{uid}");
-    }
+    public async Task DeleteDomain(Guid uid) => await _httpClient.Delete($"v1/domains/{uid}");
+    
 }
